# swagger.yml
swagger: "2.0"
info:
  title: LLM Typescript Codebase Embeddings AI Search API
  description: API documentation for AI search
  version: 1.0.0
definitions: {}
paths:
  /ai-search:
    get:
      description: Search using vector db search results with AI
      tags:
        - search
      parameters:
        - type: string
          description: Text to search
          required: false
          in: query
          name: searchText
        - type: integer
          description: Limit of results from the database
          default: 3
          required: false
          in: query
          name: dbResultLimit
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              answer:
                type: string
                description: Answer to the search query
  /db-search:
    get:
      description: Search using vector db search results
      tags:
        - search
      parameters:
        - type: string
          description: Text to search
          required: false
          in: query
          name: searchText
        - type: integer
          description: Limit of results from the database
          default: 3
          required: false
          in: query
          name: dbResultLimit
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              items:
                type: array
                description: Search Results
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: ID of the search result
                    content:
                      type: string
                      description: Text of the search result
                    path:
                      type: string
                      description: File path of the search result
                    functionOrClassName:
                      type: string
                      description: The name of the function/class etc
                    type:
                      type: string
                      description: "The type of chunk, ex: function/class/interface"
  /files:
    get:
      summary: Retrieves a list of all document file names
      description: Get all document file names
      tags:
        - documents
      responses:
        "200":
          description: Default Response
          schema:
            type: array
            items:
              type: string
              description: File name
  /files/{fileName}:
    get:
      summary: Retrieves the content of a document by file name
      description: Get the text content of a document
      tags:
        - documents
      parameters:
        - type: string
          description: The name of the file to retrieve
          required: true
          in: path
          name: fileName
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              content:
                type: string
                description: The text content of the file
    delete:
      summary: Deletes a document from the server
      description: Delete a document by file name
      tags:
        - documents
      parameters:
        - type: string
          description: The name of the file to delete
          required: true
          in: path
          name: fileName
      responses:
        "200":
          description: Default Response
host: localhost:3000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
externalDocs:
  url: https://github.com/paulb896/llm-typescript-codebase-embeddings
  description: Parse and Load Typescript Codebase into Postges Vector DB which can
    be used as part of a RAG setup.